+++ api/echo.js
@@ -1,33 +1,39 @@
-import { createHandler } from '@shgysk8zer0/lambda-http/handler.js';
-import { FORM_MULTIPART, FORM_URL_ENCODED, JSON as JSON_MIME, TEXT } from '@shgysk8zer0/lambda-http/mimes.js';
+// import '@shgysk8zer0/lambda-http/polyfills.js';
+// import { createHandler } from '@shgysk8zer0/lambda-http/handler.js';
+// import { FORM_MULTIPART, FORM_URL_ENCODED, JSON as JSON_MIME, TEXT } from '@shgysk8zer0/lambda-http/mimes.js';
 
-Headers.prototype.toJSON = function() {
-	return Object.fromEntries(this);
-};
-
-Blob.prototype.toJSON = function() {
-	return { size: this.size, type: this.type };
-};
-
-FormData.prototype.toJSON = function() {
-	return Object.fromEntries(this.entries());
-};
-
-File.prototype.toJSON = function() {
-	return { name: this.name, type: this.type, size: this.size };
-};
+import { createHandler, MIME } from '@shgysk8zer0/lambda-http/lambda-http.mjs';
 
 async function getBody(req) {
-	if (req.method !== 'GET' && req.method !== 'DELETE') {
+	Headers.prototype.toJSON = function() {
+		return Object.fromEntries(this);
+	};
+
+	Blob.prototype.toJSON = function() {
+		return { size: this.size, type: this.type };
+	};
+
+	FormData.prototype.toJSON = function() {
+		return Object.fromEntries(Array.from(
+			this.keys(),
+			key => [key, this.getAll(key)]
+		));
+	};
+
+	File.prototype.toJSON = function() {
+		return { name: this.name, type: this.type, size: this.size };
+	};
+	if (req.method !== 'GET' && req.method !== 'DELETE' && req.method !== 'HEAD') {
 		switch (req.headers.get('Content-Type')?.split(';')?.at(0)?.toLowerCase()?.trim()) {
-			case JSON_MIME:
+			case MIME.JSON:
+			case 'text/json':
 				return await req.json();
 
-			case FORM_MULTIPART:
-			case FORM_URL_ENCODED:
+			case MIME.FORM_MULTIPART:
+			case MIME.FORM_URL_ENCODED:
 				return await req.formData();
 
-			case TEXT:
+			case MIME.TEXT:
 				return req.text();
 
 			default:
@@ -53,6 +59,8 @@ async function createResponse(req) {
 		method: req.method,
 		headers: Object.fromEntries(req.headers),
 		body: await getBody(req),
+	}, {
+		headers: new Headers({ 'Access-Control-Allow-Origin': 'http://locahost:9999' })
 	});
 }