+++ utils.js
@@ -13,3 +13,32 @@ export function isFormDataRequest(req) {
 	const type = getContentType(req);
 	return type !== '' && [FORM_MULTIPART, FORM_URL_ENCODED].some(mime => type.startsWith(mime));
 }
+
+export function getOriginOrReferrer(req) {
+	if (! (req instanceof Request)) {
+		return null;
+	} else if (req.headers.has('Origin')) {
+		return URL.parse('/', req.headers.get('Origin'));
+	} else if (req.headers.has('Referer')) {
+		return URL.parse('/', req.headers.get('Referer'));
+	} else {
+		return null;
+	}
+}
+
+export function isSameOriginRequest(req) {
+	if (! (req instanceof Request)) {
+		return false;
+	} else if (! URL.canParse(req.url)) {
+		console.error('Invalid URL');
+		return false;
+	} else {
+		const url = new URL(req.url);
+		const reqOrigin = getOriginOrReferrer(req);
+		return reqOrigin instanceof URL && url.origin === reqOrigin.origin;
+	}
+}
+
+export function isCORSRequest(req) {
+	return req instanceof Request && req.headers.has('Origin') && ! isSameOriginRequest(req);
+}